timeout bug fix

+++ src/server_config.rs
@@ -10,7 +10,7 @@ pub struct FerrumServerConfig {
     pub key: Option<PathBuf>,
     pub cert: Option<PathBuf>,
     pub connect_timeout: u64,
-    pub idle_timeout: u32,
+    pub idle_timeout: u64,
     pub gateway_id: String,
     pub redis_host: String,
     pub redis_user: Option<String>,
